requirements.txt
streamlit==1.23.1
beautifulsoup4==4.13.3
openai==1.39.0
PyPDF2==3.0.1
python-dotenv==0.21.1
requests==2.31.0
sentence-transformers==2.2.2
urllib3<2.0.0
